fmt: Define GLOBALS 'width' as long, not int
This change fixes fmt toy for Linux/s390x.
Without that change width value is always 0 on that arch.

+++ toys/other/fmt.c
@@ -27,7 +27,7 @@ config FMT
 #include "toys.h"
 
 GLOBALS(
-  int width;
+  long width;
 
   int level, pos;
 )
